requirements.txt
contourpy==1.3.3
cycler==0.12.1
Cython==3.2.1
dill==0.4.0
fonttools==4.60.1
kiwisolver==1.4.9
llvmlite==0.45.1
matplotlib==3.10.7
narwhals==2.11.0
numba==0.62.1
numpy==2.3.4
numpy-stl==3.2.0
numpy-typing-compat==20250818.2.3
optype==0.14.0
packaging==25.0
pillow==12.0.0
plotly==6.4.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-utils==3.9.1
scipy==1.16.3
scipy-stubs==1.16.3.0
six==1.17.0
tqdm==4.67.1
typing_extensions==4.15.0
wheel==0.46.1
